+++ packages/dice-geometry/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dice-o-rolla/dice-geometry
 
+## 0.3.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @dice-o-rolla/dice-core@0.3.0
+
 ## 0.2.0
 
 ### Patch Changes